adding authentication

+++ main.py
@@ -2,6 +2,7 @@ import os
 import logging
 from flask import Flask, request, jsonify
 from google.cloud import firestore
+from adpipsvcfuncs import fetch_gcp_secret
 
 root_logger = logging.getLogger()
 root_logger.setLevel(logging.INFO)  # Capture DEBUG, INFO, WARNING, ERROR, CRITICAL
@@ -19,8 +20,24 @@ logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)  # Capture DEBUG, INFO, WARNING, ERROR, CRITICAL
 
 app = Flask(__name__)
+
 db = firestore.Client()
-logger.info(f"Firestore client created successfully db={db._database_string}")
+if db is None:
+    logger.error(f"Error: Unable to connect to firestore database")
+    raise Exception("Error: Unable to connect to firestore database")
+logger.info(f"Connected to firestore database db: {db}")
+
+# Fetching API key from GCP secret manager
+api_key = fetch_gcp_secret('adaptive-pipeline-API-token')
+
+def check_api_key(request):
+    token = request.headers.get('Authorization')
+    IP_address = request.remote_addr
+    if token != api_key:
+        logger.error(f"Error: Unauthorized access from IP: {IP_address}")
+        return jsonify({"success": False, "error": "Not found"}), 404
+    return None
+    
 
 @app.route("/")
 def service_working_confirmation():        
@@ -31,6 +48,9 @@ def service_working_confirmation():
 def create():
     logger.info(f"Recieved create request with data: {request.json}")
     try:
+        response = check_api_key(request)
+        if response:
+            return response
         data = request.json
         ref = db.collection('adaptive-pipelines').add(data)
         logger.info(f"Document created with ID: {ref[1].id}")
@@ -44,6 +64,9 @@ def create():
 def read(doc_id):
     logger.info(f"Recieved read request for doc_id: {doc_id}")
     try:
+        response = check_api_key(request)
+        if response:
+            return response
         doc = db.collection('adaptive-pipelines').document(doc_id).get()
         if doc.exists:
             logger.info(f"Document found: {doc.to_dict()}")
@@ -60,6 +83,9 @@ def read(doc_id):
 def update(doc_id):
     logger.info(f"Recieved update request for doc_id: {doc_id} with data: {request.json}")
     try:
+        response = check_api_key(request)
+        if response:
+            return response
         data = request.json
         db.collection('adaptive-pipelines').document(doc_id).update(data)
         logger.info(f"Document updated successfully")
@@ -72,6 +98,10 @@ def update(doc_id):
 @app.route('/delete/<doc_id>', methods=['DELETE'])
 def delete(doc_id):
     try:
+        logger.info(f"Recieved delete request for doc_id: {doc_id}")
+        response = check_api_key(request)
+        if response:
+            return response
         db.collection('adaptive-pipelines').document(doc_id).delete()
         return jsonify({"success": True}), 200
     except Exception as e:
